First commited as before_19389_rev1

+++ lttng/org.eclipse.linuxtools.ctf.core/src/org/eclipse/linuxtools/ctf/core/event/types/Definition.java
@@ -15,7 +15,7 @@ package org.eclipse.linuxtools.ctf.core.event.types;
 import org.eclipse.linuxtools.ctf.core.event.io.BitBuffer;
 
 /**
- * A CTF definition
+ * A CTF definiton
  *
  * A definition is like an object of a declaration class. It fills the
  * declaration with values. <br>
@@ -136,12 +136,13 @@ public abstract class Definition {
      *            The declaration which has an alignment
      * @since 2.2
      */
-    protected static void alignRead(BitBuffer input, IDeclaration declaration){
-        long mask = declaration.getAlignment() -1;
+    protected static void alignRead(BitBuffer input, IDeclaration declaration) {
+        int mask = (int) declaration.getAlignment() - 1;
+
         /*
          * The alignment is a power of 2
          */
-        long pos = input.position();
+        int pos = input.position();
         if ((pos & mask) == 0) {
             return;
         }
